requirements.txt
pywin32
pillow
pyautogui
keyboard
mouse
